Implement execute data with SWR

+++ practice-3/products-management-app/src/hooks/useProducts.ts
@@ -1,23 +1,19 @@
 // Library
+import { useEffect } from 'react';
 import useSWR from 'swr';
+import { AxiosError } from 'axios';
 
 // Constants
-import { ERROR_MESSAGES, SUCCESS_MESSAGES } from '@constants/messages';
-import { ProductTypes, FilterOrderOptions } from '@constants/index';
+import {
+  ProductTypes,
+  FilterOrderOptions,
+  ERROR_MESSAGES,
+  SUCCESS_MESSAGES,
+  URL_PRODUCTS
+} from '@constants/index';
 
 // Model
 import { Product } from '@models/product';
-
-// Service
-import {
-  addNewProduct,
-  filterProductsByTypeAndPrice,
-  getAllProduct,
-  getProduct,
-  removeProduct,
-  searchProducts,
-  updateProduct
-} from '@services/product.service';
 
 // Store
 import {
@@ -34,13 +30,11 @@ import {
   getProductsSuccess,
   getProductsFailed,
   filterProductsRequest,
-  getProductRequest,
-  getProductSuccess,
-  getProductFailed
+  addProductRequest
 } from '@store/index';
-import { URL_PRODUCTS } from '@constants/api';
-import { get } from '@helpers/clientRequests';
-import { useEffect } from 'react';
+
+// Helpers
+import { get, post, remove, update } from '@helpers/clientRequests';
 
 /**
  * This hook help execute product data
@@ -48,13 +42,17 @@ import { useEffect } from 'react';
  * @returns Actions
  */
 const useProducts = () => {
+  const { data, error, isValidating, mutate } = useSWR<Product[]>(
+    [URL_PRODUCTS],
+    get
+  );
+
   const { dispatch } = useStore();
 
-  const { data, error } = useSWR<Product[]>(URL_PRODUCTS, get);
-
+  // Handle get all products when loaded
   useEffect(() => {
-    getProducts();
-  }, [data]);
+    !isValidating && getProducts();
+  }, [isValidating]);
 
   /**
    * Get all products and save to local
@@ -62,38 +60,24 @@ const useProducts = () => {
   const getProducts = async (): Promise<void> => {
     try {
       dispatch(getProductsRequest());
-      console.log('data', data);
 
+      // Execute data when loaded
       const products: Product[] = data || [];
 
-      if (!products.length) {
+      if (error instanceof AxiosError && error.response?.status === 404) {
+        throw new Error(ERROR_MESSAGES.FILE_NOT_FOUND);
+      }
+
+      if (error instanceof AxiosError) {
         throw new Error(ERROR_MESSAGES.SERVER_RESPONSE_ERROR);
       }
 
-      dispatch(getProductsSuccess({ products }));
+      if (!error && products) {
+        dispatch(getProductsSuccess({ products }));
+      }
     } catch (error) {
       if (error instanceof Error) {
         dispatch(getProductsFailed({ errorMessage: error.message }));
-      }
-    }
-  };
-
-  /**
-   * Get Product by Id
-   *
-   * @param id string
-   * @returns Product
-   */
-  const getProductById = async (id: string): Promise<void> => {
-    try {
-      dispatch(getProductRequest());
-
-      const product: Product = await getProduct(id);
-
-      dispatch(getProductSuccess({ product }));
-    } catch (error) {
-      if (error instanceof Error) {
-        dispatch(getProductFailed({ errorMessage: error.message }));
       }
     }
   };
@@ -106,11 +90,15 @@ const useProducts = () => {
    */
   const createProduct = async (product: Product): Promise<void> => {
     try {
-      const newProduct: Product = await addNewProduct(product);
+      dispatch(addProductRequest());
+
+      const newProduct: Product = await post(URL_PRODUCTS, product);
 
       if (!newProduct) {
-        throw new Error(ERROR_MESSAGES.ADD_PRODUCT_FAILED);
+        throw new Error(ERROR_MESSAGES.SERVER_RESPONSE_ERROR);
       }
+
+      await mutate([...(data || []), newProduct], false);
 
       dispatch(
         addProductSuccess({
@@ -121,7 +109,6 @@ const useProducts = () => {
     } catch (error) {
       if (error instanceof Error) {
         dispatch(addProductFailed({ errorMessage: error.message }));
-        return;
       }
     }
   };
@@ -134,10 +121,23 @@ const useProducts = () => {
    */
   const editProduct = async (product: Product): Promise<void> => {
     try {
-      const updatedProduct: Product = await updateProduct(product);
-      if (!updatedProduct) {
+      const updatedProduct: Product = await update(
+        `${URL_PRODUCTS}/${product.id}`,
+        product
+      );
+
+      const updatedProductIndex: number =
+        data?.findIndex(
+          (product: Product) => product.id === updatedProduct.id
+        ) || 0;
+
+      if (!updatedProduct || !updatedProductIndex) {
         throw new Error(ERROR_MESSAGES.EDIT_PRODUCT_FAILED);
       }
+
+      data?.splice(updatedProductIndex, 1, updatedProduct);
+
+      await mutate([...(data || [])], false);
 
       dispatch(
         editProductSuccess({
@@ -161,11 +161,22 @@ const useProducts = () => {
    */
   const deleteProduct = async (id: string): Promise<void> => {
     try {
-      const deletedProduct: Product = await removeProduct(id);
+      const updatedProducts: Product[] = [...(data || [])];
 
-      if (!deletedProduct) {
-        throw new Error(ERROR_MESSAGES.SERVER_RESPONSE_ERROR);
+      const deletedProduct: Product = await remove(`${URL_PRODUCTS}/${id}`);
+
+      const deletedProductIndex: number =
+        updatedProducts?.findIndex(
+          (product: Product) => product.id === deletedProduct.id
+        ) || 0;
+
+      if (!deletedProduct || !deletedProductIndex) {
+        throw new Error(ERROR_MESSAGES.REMOVE_PRODUCT_FAILED);
       }
+
+      updatedProducts?.splice(deletedProductIndex, 1);
+
+      await mutate([...(updatedProducts || [])], false);
 
       dispatch(
         deleteProductSuccess({
@@ -187,7 +198,9 @@ const useProducts = () => {
    * @param input string
    */
   const searchingProducts = async (input: string) => {
-    const filterProducts: Product[] = await searchProducts(input);
+    const filterProducts: Product[] = await await get(
+      `${URL_PRODUCTS}?search=${input}`
+    );
     dispatch(searchProductsSuccess({ filterProducts, input }));
   };
 
@@ -201,9 +214,16 @@ const useProducts = () => {
     currentFilterTypeParam: ProductTypes,
     currentFilterPriceParam: FilterOrderOptions
   ) => {
-    const filteredProducts: Product[] = await filterProductsByTypeAndPrice(
-      currentFilterTypeParam,
-      currentFilterPriceParam
+    dispatch(filterProductsRequest());
+
+    const filteredProducts: Product[] = await get(
+      `${URL_PRODUCTS}?${`type=${currentFilterTypeParam}`}${
+        currentFilterPriceParam ? '&' : ''
+      }${
+        currentFilterPriceParam
+          ? `sortBy=price&order=${currentFilterPriceParam}`
+          : ''
+      }`
     );
 
     dispatch(filterProductsRequest());
@@ -219,13 +239,11 @@ const useProducts = () => {
 
   return {
     getProducts,
-    getProductById,
     deleteProduct,
     createProduct,
     editProduct,
     searchingProducts,
-    filterProducts,
-    data
+    filterProducts
   };
 };
